+++ src/index.ts
@@ -28,6 +28,7 @@ export * from "./components/Dot";
 export * from "./components/TruthIndicator";
 export * from "./components/QuickFilter";
 export * from "./components/MultiSelectFilter";
+export * from "./components/FilterBar";
 export * from "./components/ChartHeader";
 export * from "./components/Sparkline";
 export * from "./components/HeartbeatSparkline";